Stop mechanize from crashing with an AttributeError when it selects a form with a SubmitControl that has no value and is also disabled.
When mechanize encounters a SubmitControl with no value (such as an ImageControl) it attempts to set the value to a blank string. This fails if the control is disabled. This patch temporarily sets disabled to False in this situation and then immediately resets it to True after the blank string has been assigned to value.